move hover tooltip handling to hoverhandler
